backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==25.1
psycopg[binary]==3.2.5
python-dotenv==1.0.1
requests==2.32.3
google-auth==2.38.0
# scikit-learn==1.6.1
# statsmodels==0.14.4
pandas==2.2.3
numpy==2.2.2
yfinance==0.2.54
# transformers==4.48.2
# mlflow==2.19.0
celery==5.4.0
redis==5.0.1
django-celery-beat==2.7.0
gunicorn==23.0.0
pytest==8.3.4
pytest-django==4.10.0
